refactor(directory): rename `interfaces` directory to `interface`
